requirements.txt
colour~=0.1.5
ipython~=8.16.1
isosurfaces~=0.1.0
fontTools~=4.43.1
manimpango>=0.4.0.post0,<0.5.0
mapbox-earcut
matplotlib
moderngl~=5.8.2
moderngl_window
numpy~=1.26.1
Pillow~=10.0.1
pydub~=0.25.1
pygments~=2.16.1
PyOpenGL
pyperclip~=1.8.2
pyrr~=0.10.3
pyyaml~=6.0.1
rich~=13.6.0
scipy~=1.11.3
screeninfo~=0.8.1
skia-pathops
svgelements>=1.8.1
sympy~=1.12
tqdm~=4.66.1
typing-extensions; python_version < "3.11"
validators~=0.22.0

pyglet~=2.0.9
setuptools~=68.2.0
pandas~=2.1.1